moving interfaces around which complicated import cycles
